feat: Refactor page route to use dynamic id
The page route and its server component have been refactored to use a
dynamic id. The route file has been renamed to [id]/+page.svelte and
[id]/+page.server.ts respectively. The load function in the server
component now takes a RequestEvent parameter to get the id from the
params. The Tweet component is now rendered dynamically based on the
id.