feat: add await-to-js for error handling
This commit introduces the await-to-js library to the project. It is used
to handle errors in a more elegant way by using array destructuring. This
change is applied in the `getTweet` function in `+page.server.ts` and
`EmbeddedTweet.svelte`. The package.json and pnpm-lock.yaml files are
updated to include the new dependency.