refactor: remove custom component support
This commit removes the support for custom components in the Twitter
theme. The `AvatarImg` and `MediaImg` components are now directly used
instead of being passed as props. This simplifies the codebase and
removes the need for the `TwitterComponents` type. The `MediaImg`
component and `types.ts` file have been deleted as they are no longer
needed.