feat: Update package name and requirements
This commit updates the package name from 'sveltekit-tweet' to
'@ryoppippi/sveltekit-tweet'. It also adds a 'Requirements' section
to the README, specifying the required versions of SvelteKit and
Svelte. The usage instructions are updated to reflect these changes.
The 'Tweet' component is renamed to 'SvelteTweet', and the way 'data'
is accessed in the Svelte component is updated.