feat: Convert package from SvelteKit to Svelte
This commit changes the package from a SvelteKit specific package to a
general Svelte package. The package name is changed from
"@ryoppippi/sveltekit-tweet" to "svelte-tweet". The README is updated
to reflect these changes and to provide usage instructions for both
SvelteKit and Svelte. The import statements in the code examples are
also updated to use the new package name.