feat: Refactor to use local utils instead of react-tweet
This commit refactors the code to use local utility functions and types
instead of those from the 'react-tweet' package. This change is spread
across multiple files, mainly in the 'src/lib/components' directory and
involves updating import statements and references. A new 'utils.ts' file
is also added to the 'src/lib' directory which contains the utility
functions and types. The 'vite.config.ts' file is also updated to remove
the 'ssr.noExternal' configuration for 'react-tweet'.