feat(api): Replace react-tweet with custom Twitter API
This commit replaces the react-tweet library with a custom Twitter API.
The new API includes fetching and handling tweets, and it also provides
types for tweets, users, media, and more. The API is used in the
SvelteTweet component and its child components. This change makes the
package more lightweight and reduces the dependency on third-party
libraries.