refactor: Move react-tweet related files to src/lib/react-tweet
This commit moves all the react-tweet related files to a new directory
src/lib/react-tweet. This includes the API types, utility functions, and
CSS modules. The import paths in the Svelte components are updated to
reflect this change. This refactoring helps in better organization of
the codebase and separation of concerns.